New-folder/New folder (2)/requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-cors==3.0.10
requests==2.26.0
redis==4.3.4
python-dotenv==0.19.0
unidecode==1.2.0
rapidfuzz==2.0.0 